Flatten pages and components folder by placing CSS files under global css folder, and use barrel pattern to better organize files under pages and component inner folders
