+++ frontend/src/components/Collaboration/index.js
@@ -1,0 +1,3 @@
+export { default as Chat } from './Chat';
+export { default as CodeEditor } from './CodeEditor';
+export { default as SlidingPanel } from './SlidingPanel';
