+++ frontend/src/components/Question/QuestionList.js
@@ -3,11 +3,11 @@ import { useNavigate } from 'react-router-dom';
 import $ from 'jquery';
 import 'datatables.net';
 import 'datatables.net-bs5/css/dataTables.bootstrap5.min.css';
-import MatchingModal from '../MatchMaking/MatchingModal.js';
-import Spinner from '../Spinner.js';
-import { getQuestions } from '../../api/QuestionApi.js';
-import { errorHandler } from '../../utils/errors.js';
-import { getIsMaintainer, getCookie } from '../../utils/helpers.js';
+import Spinner from '../Spinner';
+import MatchingModal from '../MatchMaking/MatchingModal';
+import { getQuestions } from '../../api/QuestionApi';
+import { errorHandler } from '../../utils/errors';
+import { getIsMaintainer, getCookie } from '../../utils/helpers';
 
 const QuestionList = () => {
   const [isLoading, setIsLoading] = useState(true);
