+++ frontend/src/components/Question/index.js
@@ -1,0 +1,3 @@
+export { default as QuestionList } from './QuestionList';
+export { default as QuestionContent } from './QuestionContent';
+export { default as QuestionForm } from './QuestionForm';
