+++ frontend/src/components/User/index.js
@@ -1,0 +1,3 @@
+export { default as UserList } from './UserList';
+export { default as EditUser } from './EditUser';
+export { ViewUserTopPane, ViewUserBottomPane } from './ViewUser';
