+++ frontend/src/css/Collaboration.css
