+++ frontend/src/css/ConfirmationWindows.css
