+++ frontend/src/css/SlidingPanel.css
