+++ frontend/src/pages/PageNotFound.js
@@ -1,11 +1,11 @@
 import React from 'react';
-import logo from '../../images/logo.png';
-import './PageNotFound.css';
+import logo from '../images/logo.png';
+import '../css/PageNotFound.css';
 
 const PageNotFound = () => {
   return (
     <div className='page'>
-      <img src={logo} alt='Page Not Found' width='30%'/>
+      <img src={logo} alt='Page Not Found' width='30%' />
       <h4>Error 404</h4>
       <h3>Page Not Found</h3>
     </div>
