--- frontend/src/pages/Question/Question.js
@@ -1,5 +1,0 @@
-import CreateQuestion from './CreateQuestion';
-import EditQuestion from './EditQuestion';
-import QuestionDescription from './QuestionDescription/QuestionDescription';
-
-export { CreateQuestion, EditQuestion, QuestionDescription };
