+++ frontend/src/pages/Question/index.js
@@ -1,0 +1,3 @@
+export { default as CreateQuestion } from './CreateQuestion';
+export { default as EditQuestion } from './EditQuestion';
+export { default as QuestionDescription } from './QuestionDescription';
