+++ frontend/src/pages/User/UserProfile/ManageUserProfile.js
@@ -2,12 +2,9 @@ import React, { useEffect, useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 import Header from '../../../components/Header';
 import Spinner from '../../../components/Spinner';
-import {
-  ViewUserTopPane,
-  ViewUserBottomPane,
-} from '../../../components/User/ViewUser';
-import { getUser } from '../../../api/UserApi.js';
-import { errorHandler } from '../../../utils/errors.js';
+import { ViewUserTopPane, ViewUserBottomPane } from '../../../components/User';
+import { getUser } from '../../../api/UserApi';
+import { errorHandler } from '../../../utils/errors';
 import { Grid, Container } from '@mui/material';
 import { getCookie, getUserId } from '../../../utils/helpers';
