--- frontend/src/pages/User/UserProfile/UserProfile.js
@@ -1,5 +1,0 @@
-import ChangeUserPassword from './ChangeUserPassword';
-import NormalEditUser from './NormalEditUser';
-import ManageUserProfile from './ManageUserProfile';
-
-export { ChangeUserPassword, NormalEditUser, ManageUserProfile };
