+++ frontend/src/pages/User/UserProfile/index.js
@@ -1,0 +1,3 @@
+export { default as ChangeUserPassword } from './ChangeUserPassword';
+export { default as NormalEditUser } from './NormalEditUser';
+export { default as ManageUserProfile } from './ManageUserProfile';
