+++ frontend/src/pages/User/UsersManagement/ManageUserProfiles.js
@@ -1,7 +1,7 @@
 import React from 'react';
-import Header from '../../../../components/Header';
-import UserList from '../../../../components/User/UserList/UserList';
-import './ManageUserProfiles.css';
+import Header from '../../../components/Header';
+import { UserList } from '../../../components/User';
+import '../../../css/ManageUserProfiles.css';
 
 function ManageUserProfiles() {
   return (
