--- frontend/src/pages/User/UsersManagement/UsersManagement.js
@@ -1,5 +1,0 @@
-import AddUser from './AddUser';
-import MaintainerEditUser from './MaintainerEditUser';
-import ManageUserProfiles from './ManageUserProfiles/ManageUserProfiles';
-
-export { AddUser, MaintainerEditUser, ManageUserProfiles };
