+++ frontend/src/pages/User/UsersManagement/index.js
@@ -1,0 +1,3 @@
+export { default as AddUser } from './AddUser';
+export { default as MaintainerEditUser } from './MaintainerEditUser';
+export { default as ManageUserProfiles } from './ManageUserProfiles';
